requirements.txt
python-dotenv==1.0.1
requests==2.31.0
python-telegram-bot==13.13
google-cloud-dialogflow==2.38.0
vk-api=11.9.9